+++ custom_components/recalbox/__init__.py
@@ -10,6 +10,7 @@ from .api import RecalboxAPI
 from .intent import async_setup_intents # Pour charger les phrases Assist
 from .frontend import JSModuleRegistration
 from .translations import RecalboxTranslator
+from .custom_sentences_installer import install_sentences
 import os
 import shutil
 import logging
@@ -96,9 +97,11 @@ async def async_setup(hass: HomeAssistant, config: dict) -> bool:
 
     # Etape préliminaire :
     # Installer les phrases Assist automatiquement
-    hass.data[DOMAIN]["global"]["needs_restart"] = await async_install_sentences(hass)
+    hass.data[DOMAIN]["global"]["needs_restart"]= await hass.async_add_executor_job(
+        install_sentences, hass, DOMAIN
+    )
 
-    # enregistrement du chemin statique
+    # enregistrement du chemin statique pour la custom card
     await hass.http.async_register_static_paths([
         StaticPathConfig(
             "/recalbox",
@@ -126,76 +129,5 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
 
 
-
-
-
-# Tools : installer les custom_sentences
-
-
-
-def get_file_hash(filename):
-    """Calcule le hash MD5 d'un fichier."""
-    hash_md5 = hashlib.md5()
-    try:
-        with open(filename, "r", encoding="utf-8", newline=None) as f:
-            for line in f:
-                # On encode chaque ligne en utf-8 pour le hash
-                hash_md5.update(line.encode("utf-8"))
-        hashValue = hash_md5.hexdigest()
-        _LOGGER.debug("The file %s hash is %s", filename, hashValue)
-        return hashValue
-    except FileNotFoundError:
-        _LOGGER.debug("The file %s doesnt exist", filename)
-        return None
-
-
-async def async_install_sentences(hass: HomeAssistant) -> bool :
-    """Copie récursivement les sentences du composant vers le dossier système de HA."""
-    # Chemin source : /config/custom_components/recalbox/sentences
-    source_root = hass.config.path("custom_components", DOMAIN, "custom_sentences")
-    # Chemin destination : /config/custom_sentences
-    dest_root = hass.config.path("custom_sentences")
-    changes_made = False
-
-    if not os.path.exists(source_root):
-        _LOGGER.warning("Dossier source des sentences introuvable : %s", source_root)
-        return False
-
-    try:
-        # On parcourt les dossiers de langues (fr, en, es...)
-        for lang_dir in os.listdir(source_root):
-            source_lang_path = os.path.join(source_root, lang_dir)
-
-            if os.path.isdir(source_lang_path):
-                dest_lang_path = os.path.join(dest_root, lang_dir)
-                os.makedirs(dest_lang_path, exist_ok=True)
-                _LOGGER.debug("Comparing folders %s and %s ...", source_lang_path, dest_lang_path)
-
-                # On copie chaque fichier YAML
-                for file_name in os.listdir(source_lang_path):
-                    if file_name.endswith(".yaml"):
-                        source_file = os.path.join(source_lang_path, file_name)
-                        dest_file = os.path.join(dest_lang_path, file_name)
-                        _LOGGER.debug("Check if should copy %s to %s ...", source_file, dest_file)
-
-                        # LOGIQUE PAR HASH
-                        source_hash = get_file_hash(source_file)
-                        dest_hash = get_file_hash(dest_file)
-
-                        # Si le fichier destination n'existe pas ou si le contenu diffère
-                        if source_hash != dest_hash:
-                            _LOGGER.debug("Hashes are different")
-                            try:
-                                shutil.copy2(source_file, dest_file)
-                                _LOGGER.info("Mise à jour phrase Assist : %s", dest_file)
-                                changes_made = True
-                            except Exception as e:
-                                _LOGGER.error("Failed to copy file to %s: %s", dest_file, e)
-                        else:
-                            _LOGGER.debug("Hashes are equals, no need to copy again this file.")
-        return changes_made
-    except Exception as e:
-        _LOGGER.error("Erreur lors de l'installation des phrases Assist : %s", e)
-        return False